# backend/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PollPulse API",
    "version": "1.0.0",
    "description": "The official API documentation for the PollPulse real-time application."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development (Docker) Server"
    }
  ],
  "paths": {
    "/api/polls": {
      "post": {
        "summary": "Create a new poll",
        "tags": [
          "Polls"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPoll"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Poll created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Poll"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "get": {
        "summary": "Get all polls",
        "tags": [
          "Polls"
        ],
        "responses": {
          "200": {
            "description": "A list of all polls",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Poll"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/polls/{id}": {
      "get": {
        "summary": "Get a single poll by ID",
        "tags": [
          "Polls"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the poll to get",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A single poll object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Poll"
                }
              }
            }
          },
          "404": {
            "description": "Poll not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "summary": "Update a poll's question",
        "tags": [ "Polls" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the poll to update",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "question": { "type": "string", "example": "Is the new question better?" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Poll updated successfully", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Poll" } } } },
          "400": { "description": "Invalid question format" },
          "404": { "description": "Poll not found" },
          "500": { "description": "Server error" }
        }
      },
      "delete": {
        "summary": "Delete a single poll by ID",
        "tags": [ "Polls" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the poll to delete",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Poll deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Poll deleted" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Poll not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
  },
    "/api/polls/{id}/vote": {
      "post": {
        "summary": "Cast a vote on a specific poll option",
        "tags": [
          "Polls"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the poll to vote on",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewVote"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vote cast successfully, returns updated poll",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Poll"
                }
              }
            }
          },
          "404": {
            "description": "Poll or Option not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PollOption": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "6735e..."
          },
          "name": {
            "type": "string",
            "example": "React"
          },
          "votes": {
            "type": "number",
            "example": 0
          }
        }
      },
      "Poll": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "6735e..."
          },
          "question": {
            "type": "string",
            "example": "Which is the best framework?"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PollOption"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "NewPoll": {
        "type": "object",
        "description": "The data required to create a new poll",
        "required": [
          "question",
          "options"
        ],
        "properties": {
          "question": {
            "type": "string",
            "example": "Which is the best framework?"
          },
          "options": {
            "type": "array",
            "description": "An array of strings for the poll options",
            "items": {
              "type": "string"
            },
            "example": [
              "Angular",
              "React",
              "Vue"
            ]
          }
        }
      },
      "NewVote": {
        "type": "object",
        "description": "The data required to cast a vote",
        "required": [
          "optionId"
        ],
        "properties": {
          "optionId": {
            "type": "string",
            "description": "The _id of the specific option to vote for",
            "example": "6735e..."
          }
        }
      }
    }
  }
}